use db for serving files
